+++ data/src/main/java/team/retum/data/remote/datasource/declaration/StudentDataSource.kt
@@ -2,7 +2,9 @@ package team.retum.data.remote.datasource.declaration
 
 import team.retum.data.remote.request.students.EditProfileImageRequest
 import team.retum.data.remote.request.students.ResetPasswordRequest
+import team.retum.data.remote.request.user.SignUpRequest
 import team.retum.data.remote.response.students.FetchStudentInformationResponse
+import team.retum.data.remote.response.user.SignUpResponse
 
 interface StudentsDataSource {
     suspend fun fetchStudentInformation(): FetchStudentInformationResponse
@@ -17,4 +19,8 @@ interface StudentsDataSource {
         gcn: Int,
         name: String,
     )
+
+    suspend fun signUp(
+        signUpRequest: SignUpRequest,
+    ): SignUpResponse
 }
