+++ data/src/main/java/team/retum/data/remote/datasource/implementation/UserDataSourceImpl.kt
@@ -44,15 +44,6 @@ class UserDataSourceImpl @Inject constructor(
             )
         }.sendRequest()
 
-    override suspend fun signUp(
-        signUpRequest: SignUpRequest,
-    ) = HttpHandler<SignUpResponse>()
-        .httpRequest {
-            userApi.signUp(
-                signUpRequest = signUpRequest,
-            )
-        }.sendRequest()
-
     override suspend fun setUserInfo(
         signInResponse: SignInResponse,
     ) {
