+++ domain/src/main/java/team/retum/domain/repository/StudentRepository.kt
@@ -4,6 +4,7 @@ import team.retum.domain.entity.student.StudentInformationEntity
 import team.retum.domain.param.students.EditProfileImageParam
 import team.retum.domain.param.students.ResetPasswordParam
 import team.retum.domain.param.user.CheckStudentExistsParam
+import team.retum.domain.param.user.SignUpParam
 
 interface StudentRepository {
     suspend fun fetchStudentInformation(): StudentInformationEntity
@@ -14,7 +15,7 @@ interface StudentRepository {
 
     suspend fun editProfileImage(editProfileImageParam: EditProfileImageParam)
 
-    suspend fun checkStudentExists(
-        checkStudentExistsParam: CheckStudentExistsParam,
-    )
+    suspend fun checkStudentExists(checkStudentExistsParam: CheckStudentExistsParam)
+
+    suspend fun signUp(signUpParam: SignUpParam)
 }
