+++ domain/src/main/java/team/retum/domain/usecase/user/SignUpUseCase.kt
@@ -1,16 +1,16 @@
 package team.retum.domain.usecase.user
 
 import team.retum.domain.param.user.SignUpParam
-import team.retum.domain.repository.UserRepository
+import team.retum.domain.repository.StudentRepository
 import javax.inject.Inject
 
 class SignUpUseCase @Inject constructor(
-    private val userRepository: UserRepository,
+    private val studentRepository: StudentRepository,
 ) {
     suspend operator fun invoke(
         signUpParam: SignUpParam,
     ) = kotlin.runCatching {
-        userRepository.signUp(
+        studentRepository.signUp(
             signUpParam = signUpParam,
         )
     }
